refactor(core): export ClassifyResult and compute_confidence from barrel file
eval.py now imports from .core (the barrel) instead of .core.protocol,
following Sliced Bread architecture rules.
